// src/app/models/data.ts
export class Data {
    id?: number;
    number?: string;
    title?:  string;
    projectName?: string;
    client?: string;
    status?: string;
    locationPath?: string;
    locationRadius?: string;
    workLocation?: string;
    primaryContact?: string;
    billable?: string;
    owningOrg?: string;
    fromLevel: any; // high number
    toLevel: any; // low number level1-5 type shit
    startDate?: string;
    createDate?: string; 
    roleDTE?: string;
    roleClearance?: string;
    locationType?: string;
    acceptingResumes?: string;
    percentOnsite?: string;
}